LayoutTests/html5lib/runner-expected-html5.txt

CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 2: PASS
CONSOLE MESSAGE: line 2: FOO<span>BAR</span>BAZ
Content-Type: text/plain
resources/tests1.dat:
30
52
53
78
80
103

Test 30 of 113 in resources/tests1.dat failed. Input:
<a><table><td><a><table></table><a></tr><a></table><b>X</b>C<a>Y
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|       <a>
|         <b>
|           "X"
|         "C"
|       <a>
|         "Y"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|     <a>
|       <b>
|         "X"
|       "C"
|     <a>
|       "Y"

Test 52 of 113 in resources/tests1.dat failed. Input:
<p id=a><b><p id=b></b>TEST
Got:
| <html>
|   <head>
|   <body>
|     <p>
|       id="a"
|       <b>
|     <p>
|       id="b"
|       <b>
|         "TEST"
Expected:
| <html>
|   <head>
|   <body>
|     <p>
|       id="a"
|       <b>
|     <p>
|       id="b"
|       "TEST"

Test 53 of 113 in resources/tests1.dat failed. Input:
<b id=a><p><b id=b></p></b>TEST
Got:
| <html>
|   <head>
|   <body>
|     <b>
|       id="a"
|       <p>
|         <b>
|           id="b"
|       <b>
|         "TEST"
Expected:
| <html>
|   <head>
|   <body>
|     <b>
|       id="a"
|       <p>
|         <b>
|           id="b"
|       "TEST"

Test 78 of 113 in resources/tests1.dat failed. Input:
<a href="blah">aba<table><a href="foo">br<tr><td></td></tr>x</table>aoe
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|       <a>
|         href="foo"
|         "br"
|       "x"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|     "aoe"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|       <a>
|         href="foo"
|         "br"
|       <a>
|         href="foo"
|         "x"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|     <a>
|       href="foo"
|       "aoe"

Test 80 of 113 in resources/tests1.dat failed. Input:
<table><a href="blah">aba<tr><td><a href="foo">br</td></tr>x</table>aoe
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|     "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <a>
|               href="foo"
|               "br"
|     "aoe"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|     <a>
|       href="blah"
|       "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <a>
|               href="foo"
|               "br"
|     <a>
|       href="blah"
|       "aoe"

Test 103 of 113 in resources/tests1.dat failed. Input:
<a><table><td><a><table></table><a></tr><a></table><a>
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|       <a>
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|     <a>
resources/tests2.dat: PASS

resources/tests3.dat: PASS

resources/tests4.dat:
3
5
6
7

Test 3 of 7 in resources/tests4.dat failed. Input:
textarea content with <em>pseudo</em> <foo>markup
Got:
| "textarea content with "
| <em>
|   "pseudo"
| " "
| <foo>
|   "markup"
Expected:
| "textarea content with <em>pseudo</em> <foo>markup"

Test 5 of 7 in resources/tests4.dat failed. Input:
</plaintext>
Got:
| 
Expected:
| "</plaintext>"

Test 6 of 7 in resources/tests4.dat failed. Input:
setting html's innerHTML
Got:
| "setting html's innerHTML"
Expected:
| <head>
| <body>
|   "setting html's innerHTML"

Test 7 of 7 in resources/tests4.dat failed. Input:
<title>setting head's innerHTML</title>
Got:
| 
Expected:
| <title>
|   "setting head's innerHTML"
resources/tests5.dat: PASS

resources/tests6.dat:
27
30
37
44
45
47

Test 27 of 51 in resources/tests6.dat failed. Input:
foo<col>
Got:
| 
Expected:
| <col>

Test 30 of 51 in resources/tests6.dat failed. Input:
</frameset><frame>
Got:
| 
Expected:
| <frame>

Test 37 of 51 in resources/tests6.dat failed. Input:
<caption><col><colgroup><tbody><tfoot><thead><tr>
Got:
| <caption>
| <colgroup>
| <tbody>
| <tfoot>
| <thead>
|   <tr>
Expected:
| <tr>

Test 44 of 51 in resources/tests6.dat failed. Input:
</table><tr>
Got:
| <tr>
Expected:
| <tbody>
|   <tr>

Test 45 of 51 in resources/tests6.dat failed. Input:
<body></body></html>
Got:
| 
Expected:
| <head>
| <body>

Test 47 of 51 in resources/tests6.dat failed. Input:
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"><html></html>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "">
| <html>
|   <head>
|   <body>
resources/tests7.dat:
24
27
30

Test 24 of 30 in resources/tests7.dat failed. Input:
<body>X</body></body>
Got:
| "X"
Expected:
| <head>
| <body>
|   "X"

Test 27 of 30 in resources/tests7.dat failed. Input:
<table><b><tr><td>aaa</td></tr>bbb</table>ccc
Got:
| <html>
|   <head>
|   <body>
|     <b>
|     "bbb"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "aaa"
|     "ccc"
Expected:
| <html>
|   <head>
|   <body>
|     <b>
|     <b>
|       "bbb"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "aaa"
|     <b>
|       "ccc"

Test 30 of 30 in resources/tests7.dat failed. Input:
<select><keygen>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
|       <option>
|         "2048 (High Grade)"
|       <option>
|         "1024 (Medium Grade)"
|       <option>
|         "512 (Low Grade)"
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
resources/tests8.dat: PASS

resources/tests9.dat:
1
2
5
6
7
8
9
10
11
12
13
14
15
18
19
22
23
24
25

Test 1 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><math></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>

Test 2 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><math></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>

Test 5 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><math></math></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|     <table>

Test 6 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><math><mi>foo</mi></math></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|     <table>

Test 7 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><math><mi>foo</mi><mi>bar</mi></math></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>

Test 8 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><math><mi>foo</mi><mi>bar</mi></math></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|     <table>
|       <tbody>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>
|       <tbody>

Test 9 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><math><mi>foo</mi><mi>bar</mi></math></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|     <table>
|       <tbody>
|         <tr>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>
|       <tbody>
|         <tr>

Test 10 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><td><math><mi>foo</mi><mi>bar</mi></math></td></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math>
|               <mi>
|                 "foo"
|               <mi>
|                 "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math math>
|               <math mi>
|                 "foo"
|               <math mi>
|                 "bar"

Test 11 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><td><math><mi>foo</mi><mi>bar</mi></math><p>baz</td></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math>
|               <mi>
|                 "foo"
|               <mi>
|                 "bar"
|             <p>
|               "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math math>
|               <math mi>
|                 "foo"
|               <math mi>
|                 "bar"
|             <p>
|               "baz"

Test 12 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi></math><p>baz</caption></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math>
|           <mi>
|             "foo"
|           <mi>
|             "bar"
|         <p>
|           "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|         <p>
|           "baz"

Test 13 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math>
|           <mi>
|             "foo"
|           <mi>
|             "bar"
|           <p>
|             "baz"
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|         <p>
|           "baz"
|     <p>
|       "quux"

Test 14 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math>
|           <mi>
|             "foo"
|           <mi>
|             "bar"
|           "baz"
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|           "baz"
|     <p>
|       "quux"

Test 15 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><colgroup><math><mi>foo</mi><mi>bar</mi><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|       <p>
|         "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"

Test 18 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body></body></html><math><mi>foo</mi><mi>bar</mi><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|       <p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"

Test 19 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body></body><math><mi>foo</mi><mi>bar</mi><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|       <p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"

Test 22 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo><math xlink:href=foo></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <math>
|       xlink:href="foo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <math math>
|       xlink href="foo"

Test 23 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo></mi></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math>
|       <mi>
|         xlink:href="foo"
|         xml:lang="en"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"

Test 24 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo /></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math>
|       <mi>
|         xlink:href="foo"
|         xml:lang="en"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"

Test 25 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo />bar</math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math>
|       <mi>
|         xlink:href="foo"
|         xml:lang="en"
|         "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"
|       "bar"
resources/tests10.dat:
13
14
15
18
19

Test 13 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|           <svg p>
|             "baz"
|             <svg p>
|               "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|         <p>
|           "baz"
|     <p>
|       "quux"

Test 14 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|           "baz"
|           <svg p>
|             "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|           "baz"
|     <p>
|       "quux"

Test 15 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><colgroup><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|       <svg p>
|         "baz"
|         <svg p>
|           "quux"
|     <table>
|       <colgroup>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"

Test 18 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body></body></html><svg><g>foo</g><g>bar</g><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|       <svg p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"

Test 19 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body></body><svg><g>foo</g><g>bar</g><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|       <svg p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"
resources/tests11.dat:
4
8

Test 4 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><math attributeName='' attributeType='' baseFrequency='' baseProfile='' calcMode='' clipPathUnits='' contentScriptType='' contentStyleType='' diffuseConstant='' edgeMode='' externalResourcesRequired='' filterRes='' filterUnits='' glyphRef='' gradientTransform='' gradientUnits='' kernelMatrix='' kernelUnitLength='' keyPoints='' keySplines='' keyTimes='' lengthAdjust='' limitingConeAngle='' markerHeight='' markerUnits='' markerWidth='' maskContentUnits='' maskUnits='' numOctaves='' pathLength='' patternContentUnits='' patternTransform='' patternUnits='' pointsAtX='' pointsAtY='' pointsAtZ='' preserveAlpha='' preserveAspectRatio='' primitiveUnits='' refX='' refY='' repeatCount='' repeatDur='' requiredExtensions='' requiredFeatures='' specularConstant='' specularExponent='' spreadMethod='' startOffset='' stdDeviation='' stitchTiles='' surfaceScale='' systemLanguage='' tableValues='' targetX='' targetY='' textLength='' viewBox='' viewTarget='' xChannelSelector='' yChannelSelector='' zoomAndPan=''></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       contentscripttype=""
|       contentstyletype=""
|       diffuseconstant=""
|       edgemode=""
|       externalresourcesrequired=""
|       filterres=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       contentscripttype=""
|       contentstyletype=""
|       diffuseconstant=""
|       edgemode=""
|       externalresourcesrequired=""
|       filterres=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""

Test 8 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><math><altGlyph /><altGlyphDef /><altGlyphItem /><animateColor /><animateMotion /><animateTransform /><clipPath /><feBlend /><feColorMatrix /><feComponentTransfer /><feComposite /><feConvolveMatrix /><feDiffuseLighting /><feDisplacementMap /><feDistantLight /><feFlood /><feFuncA /><feFuncB /><feFuncG /><feFuncR /><feGaussianBlur /><feImage /><feMerge /><feMergeNode /><feMorphology /><feOffset /><fePointLight /><feSpecularLighting /><feSpotLight /><feTile /><feTurbulence /><foreignObject /><glyphRef /><linearGradient /><radialGradient /><textPath /></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <altglyph>
|         <altglyphdef>
|           <altglyphitem>
|             <animatecolor>
|               <animatemotion>
|                 <animatetransform>
|                   <clippath>
|                     <feblend>
|                       <fecolormatrix>
|                         <fecomponenttransfer>
|                           <fecomposite>
|                             <feconvolvematrix>
|                               <fediffuselighting>
|                                 <fedisplacementmap>
|                                   <fedistantlight>
|                                     <feflood>
|                                       <fefunca>
|                                         <fefuncb>
|                                           <fefuncg>
|                                             <fefuncr>
|                                               <fegaussianblur>
|                                                 <feimage>
|                                                   <femerge>
|                                                     <femergenode>
|                                                       <femorphology>
|                                                         <feoffset>
|                                                           <fepointlight>
|                                                             <fespecularlighting>
|                                                               <fespotlight>
|                                                                 <fetile>
|                                                                   <feturbulence>
|                                                                     <foreignobject>
|                                                                       <glyphref>
|                                                                         <lineargradient>
|                                                                           <radialgradient>
|                                                                             <textpath>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math altglyph>
|       <math altglyphdef>
|       <math altglyphitem>
|       <math animatecolor>
|       <math animatemotion>
|       <math animatetransform>
|       <math clippath>
|       <math feblend>
|       <math fecolormatrix>
|       <math fecomponenttransfer>
|       <math fecomposite>
|       <math feconvolvematrix>
|       <math fediffuselighting>
|       <math fedisplacementmap>
|       <math fedistantlight>
|       <math feflood>
|       <math fefunca>
|       <math fefuncb>
|       <math fefuncg>
|       <math fefuncr>
|       <math fegaussianblur>
|       <math feimage>
|       <math femerge>
|       <math femergenode>
|       <math femorphology>
|       <math feoffset>
|       <math fepointlight>
|       <math fespecularlighting>
|       <math fespotlight>
|       <math fetile>
|       <math feturbulence>
|       <math foreignobject>
|       <math glyphref>
|       <math lineargradient>
|       <math radialgradient>
|       <math textpath>
resources/tests12.dat:
1
2

Test 1 of 2 in resources/tests12.dat failed. Input:
<!DOCTYPE html><body><p>foo<math><mtext><i>baz</i></mtext><annotation-xml><svg><desc><b>eggs</b></desc><g><foreignObject><P>spam<TABLE><tr><td><img></td></table></foreignObject></g><g>quux</g></svg></annotation-xml></math>bar
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|       <math>
|         <mtext>
|           <i>
|             "baz"
|         <annotation-xml>
|           <svg svg>
|             <svg desc>
|               <svg b>
|                 "eggs"
|             <svg g>
|               <svg foreignObject>
|                 <svg p>
|                   "spam"
|                   <svg table>
|                     <svg tr>
|                       <svg td>
|                         <svg img>
|                           <svg g>
|                             "quux"
|                           "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|       <math math>
|         <math mtext>
|           <i>
|             "baz"
|         <math annotation-xml>
|           <svg svg>
|             <svg desc>
|               <b>
|                 "eggs"
|             <svg g>
|               <svg foreignObject>
|                 <p>
|                   "spam"
|                 <table>
|                   <tbody>
|                     <tr>
|                       <td>
|                         <img>
|             <svg g>
|               "quux"
|       "bar"

Test 2 of 2 in resources/tests12.dat failed. Input:
<!DOCTYPE html><body>foo<math><mtext><i>baz</i></mtext><annotation-xml><svg><desc><b>eggs</b></desc><g><foreignObject><P>spam<TABLE><tr><td><img></td></table></foreignObject></g><g>quux</g></svg></annotation-xml></math>bar
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "foo"
|     <math>
|       <mtext>
|         <i>
|           "baz"
|       <annotation-xml>
|         <svg svg>
|           <svg desc>
|             <svg b>
|               "eggs"
|           <svg g>
|             <svg foreignObject>
|               <svg p>
|                 "spam"
|                 <svg table>
|                   <svg tr>
|                     <svg td>
|                       <svg img>
|                         <svg g>
|                           "quux"
|                         "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "foo"
|     <math math>
|       <math mtext>
|         <i>
|           "baz"
|       <math annotation-xml>
|         <svg svg>
|           <svg desc>
|             <b>
|               "eggs"
|           <svg g>
|             <svg foreignObject>
|               <p>
|                 "spam"
|               <table>
|                 <tbody>
|                   <tr>
|                     <td>
|                       <img>
|           <svg g>
|             "quux"
|     "bar"
resources/tests14.dat: PASS

resources/tests15.dat: PASS

resources/tests16.dat: PASS

resources/webkit01.dat: PASS

resources/doctype01.dat: PASS

resources/scriptdata01.dat: PASS

resources/entities01.dat:
2
5

Test 2 of 68 in resources/entities01.dat failed. Input:
FOO&gtBAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO&gtBAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO>BAR"

Test 5 of 68 in resources/entities01.dat failed. Input:
I'm &notit; I tell you
Got:
| <html>
|   <head>
|   <body>
|     "I'm &notit; I tell you"
Expected:
| <html>
|   <head>
|   <body>
|     "I'm ¬it; I tell you"
resources/entities02.dat: PASS

resources/comments01.dat: PASS

resources/adoption01.dat:
3

Test 3 of 9 in resources/adoption01.dat failed. Input:
<a>1<button>2</a>3</button>
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "23"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "2"
|     "3"
resources/inbody01.dat: PASS

resources/isindex.dat: PASS

resources/tables01.dat: PASS
#EOF
